Fix bug in the max stack calculation
When we encounter a GOTO instruction for which we haven't encountered the
label yet, then we don't know its stackDiff contribution until after the
continuation traversal has finished. That's why we reset the stackSize and
maxStack counters and compensate for this later on when we know the stackDiff
contribution of the branching instruction.